Centralized CRC32
